+++ index.html
@@ -231,6 +231,6 @@
   </template>
 <script src="js/main.js" type="module"></script>
 <script src="pristine/pristine.min.js"></script>
-<script src="/nouislider/nouislider.js"></script>
+<script src="nouislider/nouislider.js"></script>
 </body>
 </html>
