+++ js/form.js
@@ -1,4 +1,5 @@
 import {isEscapeKey} from './util.js';
+import {onFilterButtonChange, onScaleButtonClick, scaleContainer, effectList, sliderWrapper} from './filters.js';
 
 const MAX_STRING_LENGTH = 140;
 const HASHTAGS_QUANTITY = 5;
@@ -11,13 +12,6 @@ const body = document.querySelector('body');
 const hashtagsField = document.querySelector('.text__hashtags');
 const commentsField = document.querySelector('.text__description');
 const regex = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
-const scaleValue = document.querySelector('.scale__control--value');
-const scaleContainer = document.querySelector('.img-upload__scale');
-const slider = document.querySelector('.effect-level__slider');
-const effectValue = document.querySelector('.effect-level__value');
-const scaleStep = 25;
-const minScale = 25;
-const maxScale = 100;
 
 
 const checkCommentsLength = (value) => value.length <= MAX_STRING_LENGTH;
@@ -56,55 +50,15 @@ const onPopupCloseButtonClick = () => {
   closeUploadPopup ();
 };
 
-noUiSlider.create(slider, {
-  range: {
-    min: 0,
-    max: 1,
-  },
-  start: 0,
-  step: 0.1,
-  connect: 'lower',
-  format: {
-    to: (value) => {
-      if (Number.isInteger(value)) {
-        return value.toFixed(0);
-      }
-      return value.toFixed(1);
-    },
-    from: (value) => parseFloat(value),
-  },
-});
-
-slider.noUiSlider.on('update', () => {
-  effectValue.value = slider.noUiSlider.get();
-});
-
-const onScaleButtonClick = (evt) => {
-  const scaleInput = Number.parseInt(scaleValue.value, 10);
-  let scaleCount;
-  const buttonScale = evt.target;
-  if (buttonScale.matches('.scale__control--bigger') && scaleInput < maxScale) {
-    scaleCount =  scaleInput + scaleStep;
-    scaleValue.value = `${scaleCount}%`;
-  } else {
-    scaleValue.value = `${maxScale}%`;
-  }
-
-  if (buttonScale.matches('.scale__control--smaller') && scaleInput > minScale) {
-    scaleCount = scaleInput - scaleStep;
-    scaleValue.value = `${scaleCount}%`;
-  } else {
-    scaleValue.value = `${minScale}%`;
-  }
-  imgPreview.style.transform = `scale(${scaleCount / 100})`;
-};
-
 function closeUploadPopup  () {
   editPhoto.classList.add('hidden');
   body.classList.remove('modal-open');
   document.removeEventListener('keydown', onPopupEscKeydown);
   document.removeEventListener('click', onPopupCloseButtonClick);
   scaleContainer.removeEventListener('click', onScaleButtonClick);
+  effectList.removeEventListener('change', onFilterButtonChange);
+  imgPreview.removeAttribute('class');
+  imgPreview.removeAttribute('style');
   form.reset();
 }
 
@@ -132,7 +86,9 @@ function showUploadPopup (evt) {
   buttonCancel.addEventListener('click', onPopupCloseButtonClick);
   document.addEventListener('keydown',onPopupEscKeydown);
   onFocusBlurEscKeydown();
+  sliderWrapper.classList.add('hidden');
   scaleContainer.addEventListener('click', onScaleButtonClick);
+  effectList.addEventListener('change', onFilterButtonChange);
 }
 
 const validateForm = () => {
