+++ js/main.js
@@ -6,4 +6,3 @@ import {renderUploadForm} from './form.js';
 createPhotos();
 renderUserPhotos();
 renderUploadForm();
-
